This patch makes another round of fixes to improve ansi-C compatibility, this time using the open-watcom compiler. An effort was done to make the right changes to code, not just wipe out the errors or warnings. For example, if prototypes, definition and usage of functions differ, examined the purpose and uses of the function to determine the right fix.
   The partial list of fixes includes:

 - Missing prototypes of functions.
 - Assignment of variables without appropriate casting.
 - Inadequate typing of variables.
 - Differences in definition, prototypes and use.
 - Functions defined to return some value but did not return anything.
 - Definition of zero length arrays.

   Code size is unchanged.

Committed-by: Jody Lee Bruchon <jody@jodybruchon.com>

+++ elks/fs/pipe.c
@@ -79,7 +79,7 @@ char *get_pipe_mem(void)
 }
 
 
-static int pipe_read(register struct inode *inode, struct file *filp,
+static size_t pipe_read(register struct inode *inode, struct file *filp,
 		     char *buf, int count)
 {
     size_t chars = 0, size = 0, read = 0;
@@ -128,7 +128,7 @@ static int pipe_read(register struct inode *inode, struct file *filp,
     return 0;
 }
 
-static int pipe_write(register struct inode *inode, struct file *filp,
+static size_t pipe_write(register struct inode *inode, struct file *filp,
 		      char *buf, int count)
 {
     register char *pipebuf;
@@ -181,7 +181,7 @@ static int pipe_write(register struct inode *inode, struct file *filp,
     }
     inode->i_ctime = inode->i_mtime = CURRENT_TIME;
 
-    return (int) written;
+    return written;
 }
 
 static void pipe_read_release(register struct inode *inode, struct file *filp)
@@ -242,7 +242,7 @@ static int pipe_rdwr_open(register struct inode *inode,
     return 0;
 }
 
-static int bad_pipe_rw(struct inode *inode, struct file *filp,
+static size_t bad_pipe_rw(struct inode *inode, struct file *filp,
 		       char *buf, int count)
 {
     debug("PIPE: bad rw called.\n");
